models/vlms/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.45.0
accelerate>=0.21.0
pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0
regex>=2023.0.0
flash-attn>=2.0.0
sentencepiece>=0.1.99
timm>=0.9.0 